+++ DdsTextureDescription.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using DDSTextureLoader.NET.TextureParsing;
 using TerraFX.Interop;
 
 #nullable enable
@@ -33,29 +34,67 @@ namespace DDSTextureLoader.NET
             SubresourceData = subresourceData;
             AlphaMode = alphaMode;
         }
-
+        
+        private readonly Size3 _size;
         
         /// <summary>
         /// The buffer that contains the data referenced by <see cref="SubresourceData"/>
         /// </summary>
         public Memory<byte> BitData { get; }
+        
         /// <summary>
         /// The dimension of the DDS data
         /// </summary>
         public D3D12_RESOURCE_DIMENSION ResourceDimension { get; }
-        private readonly Size3 _size;
+        
         /// <summary>
-        /// The height
+        /// The height of the texture
         /// </summary>
         public uint Height => _size.Height;
+        
+        /// <summary>
+        /// The width of the texture
+        /// </summary>
         public uint Width  => _size.Width;
+        
+        /// <summary>
+        /// The depth of the texture
+        /// </summary>
         public uint Depth => _size.Depth;
+        
+        /// <summary>
+        /// The number of MIPs present
+        /// </summary>
         public uint MipCount { get; }
+        
+        /// <summary>
+        /// The size of the texture, if depth is 1
+        /// </summary>
         public uint ArraySize { get; }
+        
+        /// <summary>
+        /// The format of the texture
+        /// </summary>
         public DXGI_FORMAT Format { get; }
+        
+        /// <summary>
+        /// Whether SRGB should be forced
+        /// </summary>
         public bool ForceSrgb { get; }
+        
+        /// <summary>
+        /// Whether the texture is a cube map
+        /// </summary>
         public bool IsCubeMap { get; }
+        
+        /// <summary>
+        /// The subresource data, relative to <see cref="BitData"/>, for upload
+        /// </summary>
         public Memory<ManagedSubresourceData> SubresourceData { get; }
+        
+        /// <summary>
+        /// The alpha mode of the texture
+        /// </summary>
         public DDS_ALPHA_MODE AlphaMode { get; }
     }
 }
