+++ ManagedSubresourceData.cs
