+++ TextureParsing/D3D11_RESOURCE_MISC_FLAG.cs
@@ -1,0 +1,26 @@
+﻿using System;
+
+namespace DDSTextureLoader.NET.TextureParsing
+{
+    [Flags]
+    internal enum D3D11_RESOURCE_MISC_FLAG
+    {
+        D3D11_RESOURCE_MISC_GENERATE_MIPS,
+        D3D11_RESOURCE_MISC_SHARED,
+        D3D11_RESOURCE_MISC_TEXTURECUBE,
+        D3D11_RESOURCE_MISC_DRAWINDIRECT_ARGS,
+        D3D11_RESOURCE_MISC_BUFFER_ALLOW_RAW_VIEWS,
+        D3D11_RESOURCE_MISC_BUFFER_STRUCTURED,
+        D3D11_RESOURCE_MISC_RESOURCE_CLAMP,
+        D3D11_RESOURCE_MISC_SHARED_KEYEDMUTEX,
+        D3D11_RESOURCE_MISC_GDI_COMPATIBLE,
+        D3D11_RESOURCE_MISC_SHARED_NTHANDLE,
+        D3D11_RESOURCE_MISC_RESTRICTED_CONTENT,
+        D3D11_RESOURCE_MISC_RESTRICT_SHARED_RESOURCE,
+        D3D11_RESOURCE_MISC_RESTRICT_SHARED_RESOURCE_DRIVER,
+        D3D11_RESOURCE_MISC_GUARDED,
+        D3D11_RESOURCE_MISC_TILE_POOL,
+        D3D11_RESOURCE_MISC_TILED,
+        D3D11_RESOURCE_MISC_HW_PROTECTED
+    };
+}
