+++ TextureParsing/Size2.cs
@@ -1,0 +1,20 @@
+﻿namespace DDSTextureLoader.NET.TextureParsing
+{
+    internal struct Size2
+    {
+        public Size2(uint height, uint width)
+        {
+            Height = height;
+            Width = width;
+        }
+
+        public void Deconstruct(out uint height, out uint width)
+        {
+            height = Height;
+            width = Width;
+        }
+
+        public uint Height { get; set; }
+        public uint Width { get; set; }
+    }
+}
